+++ src/Workspaces/SharedUtilitiesAndExtensions/Workspace/Core/EmbeddedLanguages/VirtualChars/IVirtualCharLanguageService.cs
@@ -1,0 +1,12 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using Microsoft.CodeAnalysis.Host;
+
+namespace Microsoft.CodeAnalysis.EmbeddedLanguages.VirtualChars
+{
+    internal interface IVirtualCharLanguageService : IVirtualCharService, ILanguageService
+    {
+    }
+}
